+++ src/gui/CraftingPage.lua
@@ -168,10 +168,16 @@ function Self:UpdateSort(refresh)
         self.frame.RecipeList.NoResultsText:SetShown(false)
 
         self.sortJob = Promise:Async(function ()
+            -- Util:DebugProfileStart("Job > Start")
+
             Promise:GetCurrent():SetPriority(5)
 
             for i,recipeID in ipairs(recipeIDs) do
+                Util:DebugProfileSegment("Recipe > GetRecipeSchematic")
+
                 local recipe = C_TradeSkillUI.GetRecipeSchematic(recipeID, false)
+
+                Util:DebugProfileSegment("Recipe > Cache")
 
                 local key, time = cache:Key(recipe), Prices:GetRecipeScanTime(recipe)
                 if not cache:Has(key) or cache:Get(key)[1] ~= time then
@@ -181,6 +187,9 @@ function Self:UpdateSort(refresh)
                 local operation = cache:Get(key)[2]
                 if operation then
                     local value = Optimization:GetOperationValue(operation, method)
+
+                    Util:DebugProfileSegment("Recipe > Data Provider Insert")
+
                     if value and abs(value) ~= math.huge then
                         self.dataProvider:Insert({
                             recipeInfo = C_TradeSkillUI.GetRecipeInfo(recipeID),
@@ -195,10 +204,16 @@ function Self:UpdateSort(refresh)
                 Promise:YieldProgress(i, n)
             end
 
+            Util:DebugProfileSegment("Job > NoResultsText")
+
             self.frame.RecipeList.NoResultsText:SetShown(self.dataProvider:IsEmpty())
+
+            Util:DebugProfileSegment("Job > Data Provider Invalidate")
 
             -- Fix last recipe not getting sorted correctly
             self.dataProvider:Invalidate()
+
+            -- Util:DebugProfileStop()
         end):Start(function ()
             self.progressBar:Start(n)
             return function () self.progressBar:Progress(n, n) end