+++ CHANGES.md
@@ -1,3 +1,4 @@
 - Add tracking multiple orders of the same recipe
 - Properly handle finishing reagents that change resourcefulness/multicraft yield
+- Improve profit and concentration optimization performance by about 33%
 - Updated reagent, perk and recipe data